cambiado el inputText de passchange a inputPassword y cambios a las bolsas
