+++ Frontend/src/styles/client/Bags.module.css
@@ -1,25 +1,22 @@
 .container{
-    background-color: #101010;
-    height: 100vh;
+    display: grid;
+    justify-items: center;
+    grid-template-rows: auto 1fr auto ;
+    min-height: 100vh;
+    width: 100%;
 }
 
-.section1{
+
+.dashboardWidgets{
     display: flex;
     justify-content: center;
-    align-items: center;
-    width: 100%;
-    /* border-radius: 40px; */
-    /* background: linear-gradient(to right, #919191, #727272); */
-    padding: 30px 0;
-    margin-bottom: 30px;
-    .dashboardWidgets{
-        display: flex;
-        justify-content: center;
-        flex-wrap: wrap;
-        gap: 20px;
-        width: 80%;
+    flex-wrap: wrap;
+    gap: 20px;
+    width: 80%;
+    max-width: 1700px;
     }
-    }
+
+
 
 .section2{
     background-color: #4e4646;
@@ -28,5 +25,4 @@
 .gap{
     height: 7rem;
     width: 100%;
-    background-color: #101010;
 }
