+++ Frontend/src/styles/client/main.module.css
@@ -33,11 +33,6 @@
     width: 100%;
     margin-top: 30px;
     margin-bottom: 30px;
-    /* h1{
-      font-size: 40px;
-      color: #cccccc;
-      margin-bottom: 20px;
-    } */
     .dashboardWidgets{
       display: flex;
       justify-content: center;
